backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
sqlparse==0.4.4
requests==2.31.0
python-dotenv==1.0.0